// src/modules/meetings/types.ts
import { AppRouter } from "@/tRPC/routers/_app";
import { inferRouterOutputs } from "@trpc/server";

export type MeetingGetMany =
  inferRouterOutputs<AppRouter>["meetings"]["getMany"]["items"];

export type MeetingGetOne = inferRouterOutputs<AppRouter>["meetings"]["getOne"];

export enum MeetingStatus {
  Upcoming = "upcoming",
  Active = "active",
  Completed = "completed",
  Processing = "processing",
  Cancelled = "cancelled",
}


export type StreamTranscriptItem = {
  speaker_id: string;
  type : string;
  text: string;
  start_ts: string; 
  stop_ts: string; 
}